requirements-pip.txt
numpy>=1.14.3
opencv-python>=3.4.0.12
PyHamcrest>=1.9.0
pytest>=3.5.1
six>=1.11.0
pytesseract>=0.2.2
django>=2.1
celery>=4.2.0
redis>=2.10.6
djangorestframework>=3.8.2
markdown>=2.6.11
django-filter>=2.0.0c
gunicorn>=19.9.0
django-heroku>=0.3.1
Pillow>=5.3.0
pandas>=0.23.4
torchvision>=0.2.1
